requirements.txt
python-slugify
requests
beautifulsoup4
unicode
pyyaml